Initial Commit
Buggy, but working file commited after full day of working on it during
Code for Seattle 2014.  Main issue is that, while triggering action from
a HIGH signal via pushbutton or signal from pedestrain counter arduino
works, and the program always creates files on the SD card, it only
sometimes will actually write the full image data. Essentially, the
image write either works or doesn't for the entire time the pedCamera is
turned on. This may be due to voltage differences between
battery/usb/external power or due to improper "debouncing" of the
button/counter signal in the pedCamera code.